Implement configurable forwarding of all message types
This changes the default behavior: before if no topic names were provided, default
names were used. Now if a topic name is not provided for a specific message type
then messages of that message type are dropped (i.e. not forwarded to kafka).

All message types can be forwarded except messages with unknown type.

Additional statistics are provided: the number of messages processed is now split
in 3 metrics:
- Consume is the number of messages received
- Ignored is the number of messages that were not forwarded due to configuration
- Forwarded is the number of messages that were forwarded due to configuration
- note: Consume = Ignored + Forwarded

Also, ConsumeXXX submetrics are provided for each NOAA message type (LogMessage,
ContainerMetric, ...)

Note that the meaning of Consume has changed. Before it was what would now be the
sum of ConsumeLogMessage and ConsumeValueMetric. Now is the number of messages that
have been successfully consumed. Since it is now possible to configure
kafka-firehose-nozzle to drop certain type of messages, the Delay has been redefined
as Forwarded-(Publish+PublishFail)