Implement message enrichment
The enrich() function receives the original events.Envelope and and the app GUID.
It returns an Envelope that can contain additional fields (defined in ExtraContent).

Right now the only data added, if provided, is the app GUID, but in the future other
data (app|org|space) (GUID|name) can be added.